fix errors
