+++ lib/stockInfo/common.js
@@ -10,7 +10,6 @@ import {
   getAnnualizedPcnt,
   getFormattedFromToDate
 } from '@/lib/commonFunction'
-import { getPriceMADetails } from '@/lib/email/template'
 import { toAxios } from '@/lib/request'
 import { getHistoryPrice } from '@/lib/yahoo/getHistoryPrice'
 
@@ -85,39 +84,6 @@ const getFormattedHistoryPrice = async (ticker, fromDate, toDate) => {
       moment.unix(item).format('DD MMM YYYY')
     ),
     price: outputItem.indicators.quote.find(x => x).close || []
-  }
-}
-
-const conditionMatches = (id, priceMAList) => {
-  const matches = !!priceMAList.find(
-    x => x.id === id && x.tickersInfo.length > 0
-  )
-
-  return matches
-}
-
-const getStockStrength = async ticker => {
-  const genChart = false
-  const result = await getPriceMADetails({ ticker, genChart })
-  const { rs, latestHigherInputRange } = await getRelativeStrengthByDays(
-    ticker,
-    'SPY',
-    50,
-    85,
-    3
-  )
-
-  const twentyHigherFifty = conditionMatches('20>50', result.priceMAList)
-  const fiftyHigherHundredFifty = conditionMatches('50>150', result.priceMAList)
-  const fiftyLowerHundredFifty = conditionMatches('50<150', result.priceMAList)
-
-  return {
-    Ticker: result.ticker,
-    'Price Not Avail.': !result.priceAvail,
-    'RS>0': rs,
-    '85D High': latestHigherInputRange,
-    '20>50 & 50>150': twentyHigherFifty && fiftyHigherHundredFifty,
-    '20>50 & 50<150': twentyHigherFifty && fiftyLowerHundredFifty
   }
 }
 
@@ -302,7 +268,6 @@ const showHighlightQuoteDetail = (router, inputQuery) => {
 export {
   getFormattedHistoryPrice,
   getRelativeStrengthByDays,
-  getStockStrength,
   getUserBoughtList,
   getUserBoughtListDetails,
   getPriceInfo,
